+++ Changes.md
@@ -27,6 +27,7 @@ API
 - MenuButton : Added `set/getImmediate()` methods. These allow the button to execute a menu item immediately instead of showing the menu, if the menu has only a single item.
 - Signals : Added CatchingCombiner Python class, equivalent to the C++ template class.
 - FileSystemPath : Implemented Python's `os.PathLike` interface, to allow FileSystemPath objects to be passed directly to Python's standard filesystem functions.
+- PlugCreationWidget : Added standard metadata-driven widget to allow users to add plugs to nodes.
 
 Build
 -----